pyproject.toml
[project]
name = "wiki-craft"
version = "0.1.0"
description = "Document ingestion and indexing system for building wiki-style knowledge bases with full source tracking"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Wiki-Craft Team" }]
keywords = ["wiki", "document-processing", "vector-database", "knowledge-base", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Document Parsing
    "pymupdf>=1.23.0",              # PDF text extraction
    "pytesseract>=0.3.10",          # OCR for scanned PDFs
    "pillow>=10.0.0",               # Image processing for OCR
    "python-docx>=1.1.0",           # Word documents
    "openpyxl>=3.1.0",              # Excel spreadsheets
    "beautifulsoup4>=4.12.0",       # HTML parsing
    "lxml>=5.0.0",                  # XML/HTML processing
    "ebooklib>=0.18",               # EPUB parsing
    "markdown-it-py>=3.0.0",        # Markdown parsing
    "python-frontmatter>=1.1.0",    # Markdown frontmatter
    
    # ML and Embeddings
    "sentence-transformers>=2.3.0",
    "torch>=2.0.0",
    
    # Vector Database
    "chromadb>=0.4.22",
    
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",                # HTTP client for URL ingestion
    "python-magic>=0.4.27",         # File type detection
    "aiofiles>=23.2.0",             # Async file operations
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",                # For testing FastAPI
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
wiki-craft = "wiki_craft.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wiki_craft"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["wiki_craft"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
